[Step3] 코드리뷰 반영 - raceService -> carRace 클래스로 변경, 역할 분담
