+++ src/main/kotlin/step3/Main.kt
@@ -2,8 +2,10 @@ package step3
 
 import step3.car.CarRegistration
 import step3.common.CarRaceProperty
-import step3.race.RaceService
+import step3.race.CarRace
+import step3.utils.RandomNumberGenerator
 import step3.view.InputView
+import step3.view.ResultView
 
 fun main() {
     val carRaceProperty = CarRaceProperty()
@@ -14,6 +16,14 @@ fun main() {
     val carRegistration = CarRegistration(carRaceProperty)
     carRegistration.initCars()
 
-    val raceService = RaceService(carRaceProperty)
-    raceService.runRacingByRound()
+    val carRace = CarRace()
+    val resultView = ResultView(carRaceProperty)
+    for (round in CarRaceProperty.ROUND_START..carRaceProperty.roundSize) {
+        carRaceProperty.cars
+            .map {
+                carRace.moveCarPosition(it, it.canMoveForward(RandomNumberGenerator.getRandomNumber()))
+            }
+
+        resultView.printResultByRound(round)
+    }
 }
