+++ src/main/kotlin/step3/car/Car.kt
@@ -1,3 +1,9 @@
 package step3.car
 
-data class Car(var moveCount: Int)
+data class Car(var moveCount: Int) {
+    fun canMoveForward(number: Int) = number >= MOVE_CONDITION
+
+    companion object {
+        const val MOVE_CONDITION = 4
+    }
+}
