+++ src/main/kotlin/step3/race/CarRace.kt
@@ -1,0 +1,14 @@
+package step3.race
+
+import step3.car.Car
+
+class CarRace {
+
+    fun moveCarPosition(car: Car, canMoveForward: Boolean) {
+        if (canMoveForward) {
+            moveForward(car)
+        }
+    }
+
+    private fun moveForward(car: Car) = car.moveCount++
+}
